Fix incorrect computed state when remote player loads avatar faster than their wearer
When a wearer changes avatar, if a remote player loads an avatar faster than their wearer, then it is possible that the remote player will try to evaluate the avatar inventory state with an unpredictable value of the avatar for a given parameter.

This unpredictable value seems to be equal to 0 when a player joins a world. This means that if a remote player loads the avatar faster than the wearer, all of the toggles will be evaluated as being off, causing all items or avatar features to disappear.

The current attempted fix is:

- In the View, if the player is remote, then only flag for dirty checks after the `J1I7T_A_Sync` parameter is strictly greater than 127. This conditions becomes true when the avatar wearer actually driver the sync state when they have loaded the avatar and reached the "Local SetSyncValue" in the View.
- In the Transmitters, all intermediate remote transitions to a low bit state will also require  `J1I7T_A_Sync` parameter to be strictly greater than 127. This will temporarily suspend evaluation of the binary tree if the avatar wearer happened to be in the middle of changing avatar toggles again while the binary tree was being evaluated remotely.

Closes #4