Fix zero items visible for a split second
When a remote player loads an avatar, the clothes will attempt to blend as soon as the remote player receives the synced value.

However, there is a split second before the received synced value is computed into toggles, causing all the items to disappear for that split duration.

The current fix is to wait for the dirty checks to clear remotely, which is attained when the evaluation completes at the leaves of the transmitter.

Closes #6